refactor: extract PhysBone methods to separate PhysBones class
PhysBone系のメソッドをPhysBoneExtractor.csから新しいPhysBones.csファイルに分離:
- CopyVRCPhysBone()メソッドの移動
- CopyVRCPhysboneCollider()メソッドの移動
- PhysBoneExtractor.csでPhysBonesクラスを使用するよう修正
- PhysBoneとPhysBoneColliderの処理順序を適切に管理
- コードの責任分離とメンテナビリティの向上
