+++ Editor/PhysBoneExtractor.cs
@@ -4,14 +4,9 @@ using System.IO;
 using System.Linq;
 using UnityEditor;
 using UnityEngine;
-using UnityEngine.Animations;
 using VRC.Dynamics;
-using VRC.SDK3.Avatars;
-using VRC.SDK3.Avatars.Components;
-using VRC.SDK3.Dynamics.Constraint.Components;
 using VRC.SDK3.Dynamics.Contact.Components;
 using VRC.SDK3.Dynamics.PhysBone.Components;
-using YamlDotNet.Core;
 
 namespace dev.kesera2.physbone_extractor
 {
@@ -151,6 +146,7 @@ namespace dev.kesera2.physbone_extractor
                             contactSenderGameObjectName = EditorGUILayout.TextField(contactSenderGameObjectName);
                             contactReceiverGameObjectName = EditorGUILayout.TextField(contactReceiverGameObjectName);
                         }
+
                         constraintsGameObjectName = EditorGUILayout.TextField(constraintsGameObjectName);
                         vrcPositionConstraintName = EditorGUILayout.TextField(vrcPositionConstraintName);
                         vrcRotationConstraintName = EditorGUILayout.TextField(vrcRotationConstraintName);
@@ -180,15 +176,16 @@ namespace dev.kesera2.physbone_extractor
                         var constraints = new Constraints(vrcPositionConstraintName, vrcRotationConstraintName,
                             vrcParentConstraintName, vrcScaleConstraintName, vrcAimConstraintName,
                             vrcLookAtConstraintName);
-                        copiedConstraintComponents = constraints.CopyVRCConstraint(searchRoot, _avatarDynamics, constraintsGameObjectName);
-                        copiedPbColliderComponents = Utility.CopyComponent<VRCPhysBoneCollider>(searchRoot,
-                            _avatarDynamics,
-                            pbColliderGameObjectName, CopyVRCPhysboneCollider);
+                        copiedConstraintComponents =
+                            constraints.CopyVRCConstraint(searchRoot, _avatarDynamics, constraintsGameObjectName);
+                        var physBones = new PhysBones(pbGameObjectName, pbColliderGameObjectName, _isKeepPbVersion);
+                        copiedPbColliderComponents = physBones.CopyVRCPhysBoneColliders(searchRoot, _avatarDynamics);
                         copiedPbComponents =
-                            Utility.CopyComponent<VRCPhysBone>(searchRoot, _avatarDynamics, pbGameObjectName,
-                                CopyVRCPhysBone);
-                        var contacts = new Contacts(contactsGameObjectName, contactSenderGameObjectName, contactReceiverGameObjectName, splitContacts);
-                        var hasContacts = contacts.CopyVRCContacts(searchRoot, _avatarDynamics, out copiedContactSenderComponents, out copiedContactReceiverComponents);
+                            physBones.CopyVRCPhysBones(searchRoot, _avatarDynamics, copiedPbColliderComponents);
+                        var contacts = new Contacts(contactsGameObjectName, contactSenderGameObjectName,
+                            contactReceiverGameObjectName, splitContacts);
+                        var hasContacts = contacts.CopyVRCContacts(searchRoot, _avatarDynamics,
+                            out copiedContactSenderComponents, out copiedContactReceiverComponents);
                         if (!(copiedPbComponents.Count > 0 || copiedPbColliderComponents.Count > 0 || hasContacts))
                             // Destroy the GameObject of AvatarDynamics if there is no target components.
                             DestroyImmediate(_avatarDynamics);
@@ -297,106 +294,6 @@ namespace dev.kesera2.physbone_extractor
             _avatarDynamics.transform.SetParent(prefabRoot.transform);
         }
 
-
-
-
-        private void CopyVRCPhysBone(VRCPhysBone source, VRCPhysBone destination)
-        {
-            // Vesrion
-            if (_isKeepPbVersion) destination.version = source.version;
-
-            // Copy properties from source to destination
-            destination.name = source.name;
-            destination.enabled = source.enabled;
-
-            // Transforms
-            destination.rootTransform = source.rootTransform;
-            destination.ignoreTransforms = source.ignoreTransforms;
-            destination.endpointPosition = source.endpointPosition;
-            destination.multiChildType = source.multiChildType;
-
-            // Forces
-            destination.integrationType = source.integrationType;
-            destination.pull = source.pull;
-            destination.pullCurve = source.pullCurve;
-            destination.spring = source.spring;
-            destination.springCurve = source.springCurve;
-            destination.stiffness = source.stiffness;
-            destination.stiffnessCurve = source.stiffnessCurve;
-            destination.gravity = source.gravity;
-            destination.gravityCurve = source.gravityCurve;
-            destination.gravityFalloff = source.gravityFalloff;
-            destination.gravityFalloffCurve = source.gravityFalloffCurve;
-            destination.immobileType = source.immobileType;
-            destination.immobile = source.immobile;
-            destination.immobileCurve = source.immobileCurve;
-
-            // Limits
-            destination.limitType = source.limitType;
-            destination.maxAngleX = source.maxAngleX;
-            destination.maxAngleZ = source.maxAngleZ;
-            destination.maxAngleXCurve = source.maxAngleXCurve;
-            destination.maxAngleZCurve = source.maxAngleZCurve;
-            destination.staticFreezeAxis = source.staticFreezeAxis;
-            destination.limitRotation = source.limitRotation;
-            destination.limitOpacity = source.limitOpacity;
-            destination.limitRotationXCurve = source.limitRotationXCurve;
-            destination.limitRotationYCurve = source.limitRotationYCurve;
-            destination.limitRotationZCurve = source.limitRotationZCurve;
-
-            // Colliders
-            destination.radius = source.radius;
-            destination.radiusCurve = source.radiusCurve;
-            destination.allowCollision = source.allowCollision;
-            destination.colliders = new List<VRCPhysBoneColliderBase>(from sourceCollider in source.colliders
-                join copiedCollider in copiedPbColliderComponents on sourceCollider.name equals copiedCollider.name
-                select copiedCollider).ToList();
-            destination.collisionFilter = source.collisionFilter;
-
-            // Stretch & Squish
-            destination.maxStretch = source.maxStretch;
-            destination.maxStretchCurve = source.maxStretchCurve;
-            destination.stretchMotion = source.stretchMotion;
-            destination.stretchMotionCurve = source.stretchMotionCurve;
-            destination.maxSquish = source.maxSquish;
-            destination.maxSquishCurve = source.maxSquishCurve;
-
-            // Grab & Pose
-            destination.allowGrabbing = source.allowGrabbing;
-            destination.allowPosing = source.allowPosing;
-            destination.grabMovement = source.grabMovement;
-            destination.snapToHand = source.snapToHand;
-            // destination.grabFilter = source.grabFilter;
-            // destination.poseFilter = source.poseFilter;
-
-            // Options
-            destination.parameter = source.parameter;
-            destination.isAnimated = source.isAnimated;
-            destination.resetWhenDisabled = source.resetWhenDisabled;
-
-            // Gizmos
-            destination.showGizmos = source.showGizmos;
-            destination.boneOpacity = source.boneOpacity;
-            destination.limitOpacity = source.limitOpacity;
-        }
-
-        private void CopyVRCPhysboneCollider(VRCPhysBoneCollider source, VRCPhysBoneCollider destination)
-        {
-            // Copy properties from source to destination
-            destination.name = source.name;
-            destination.enabled = source.enabled;
-            destination.rootTransform = source.rootTransform;
-            destination.shapeType = source.shapeType;
-            destination.insideBounds = source.insideBounds;
-            destination.radius = source.radius;
-            destination.height = source.height;
-            destination.position = source.position;
-            destination.rotation = source.rotation;
-            destination.bonesAsSpheres = source.bonesAsSpheres;
-            destination.isGlobalCollider = source.isGlobalCollider;
-            destination.playerId = source.playerId;
-            destination.shape = source.shape;
-        }
 
         private Transform GetArmatureTransform()
         {
